requirements.txt
# SHADOWSCAN PRO - Requirements
# Developed by ROHAIB TECHNICAL | +92 306 3844400

# Core
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio>=3.4.3
httpx>=0.24.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0
playwright>=1.30.0

# Dashboard & UI
rich>=13.0.0
textual>=0.10.0
colorama>=0.4.6
asciichartpy>=1.5.0
plotext>=5.2.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.9.0

# Analysis & ML
scikit-learn>=1.2.0
joblib>=1.2.0
regex>=2023.0.0
dnspython>=2.3.0
tldextract>=3.4.0
phonenumbers>=8.13.0

# Network & Exploitation
paramiko>=3.0.0
pyOpenSSL>=23.0.0
python-nmap>=0.7.1
impacket>=0.10.0

# Export & Reporting
jinja2>=3.1.0
markdown>=3.4.0
pdfkit>=1.0.0
weasyprint>=58.0.0
openpyxl>=3.1.0
elasticsearch>=8.5.0

# Security & Anonymity
cryptography>=39.0.0
pycryptodome>=3.17.0
PySocks>=1.7.1
stem>=1.8.1
fake-useragent>=1.4.0

# API Clients
python-whois>=0.8.0
shodan>=1.28.0
github3.py>=3.2.0
tweepy>=4.12.0
praw>=7.7.0
python-gitlab>=3.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.64.0
pydantic>=1.10.0
loguru>=0.6.0
watchdog>=3.0.0
psutil>=5.9.0

# Development (optional)
pytest>=7.2.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0